make weight

+++ db/migrate/20170724052249_change_datatype_open_on_and_full_on_of_sakura.rb
@@ -1,8 +1,13 @@
 class ChangeDatatypeOpenOnAndFullOnOfSakura < ActiveRecord::Migration
   def up
-    ActiveRecord::Base.connection.execute("TRUNCATE sakuras;")
-    change_column :sakuras, :open_on, 'integer USING NULL'
-    change_column :sakuras, :full_on, 'integer USING NULL'
+    if Rails.configuration.database_configuration[Rails.env]["adapter"] == "mysql2"
+      ActiveRecord::Base.connection.execute("TRUNCATE sakuras;")
+      change_column :sakuras, :open_on, :integer
+      change_column :sakuras, :full_on, :integer
+    else
+      change_column :sakuras, :open_on, 'integer USING NULL'
+      change_column :sakuras, :full_on, 'integer USING NULL'
+    end
   end
   def down
     ActiveRecord::Base.connection.execute("TRUNCATE sakuras;")
